refactor: move some tests inside the lib
