+++ src/conversions/std/string.rs
@@ -215,6 +215,22 @@ mod tests {
     }
 
     #[test]
+    fn test_extract_str_surrogate() {
+        use crate::exceptions::PyUnicodeEncodeError;
+
+        Python::attach(|py| {
+            let value = py.eval(cr"'\ud800'", None, None).unwrap();
+            let err = value.extract::<String>().unwrap_err();
+
+            assert!(err.is_instance_of::<PyUnicodeEncodeError>(py));
+            assert_eq!(
+                err.value(py).to_string(),
+                "'utf-8' codec can't encode character '\\ud800' in position 0: surrogates not allowed"
+            );
+        });
+    }
+
+    #[test]
     fn test_extract_char() {
         Python::attach(|py| {
             let ch = '😃';
